followers-list: split components in separte files
